+++ test/fuzz/NodeFuzz.t.sol
@@ -254,12 +254,12 @@ contract NodeFuzzTest is BaseTest {
                         FEE PAYMENTS
     //////////////////////////////////////////////////////////////*/
 
-    function test_fuzz_node_payManagementFees(uint256 annualFee, uint256 protocolFee, uint256 seedAmount) public {
+    function test_fuzz_node_payManagementFees(uint64 annualFee, uint64 protocolFee, uint256 seedAmount) public {
         address ownerFeesRecipient = makeAddr("ownerFeesRecipient");
         address protocolFeesRecipient = makeAddr("protocolFeesRecipient");
 
-        annualFee = bound(annualFee, 0, 1e18);
-        protocolFee = bound(protocolFee, 0, 1e18);
+        annualFee = uint64(bound(annualFee, 0, 1e18));
+        protocolFee = uint64(bound(protocolFee, 0, 1e18));
         seedAmount = bound(seedAmount, 1e18, 1e36);
 
         vm.startPrank(owner);
@@ -286,16 +286,16 @@ contract NodeFuzzTest is BaseTest {
     }
 
     function test_fuzz_node_payManagementFees_different_durations(
-        uint256 annualFee,
-        uint256 protocolFee,
+        uint64 annualFee,
+        uint64 protocolFee,
         uint256 seedAmount,
         uint256 duration
     ) public {
         address ownerFeesRecipient = makeAddr("ownerFeesRecipient");
         address protocolFeesRecipient = makeAddr("protocolFeesRecipient");
 
-        annualFee = bound(annualFee, 0, 1e18);
-        protocolFee = bound(protocolFee, 0, 1e18);
+        annualFee = uint64(bound(annualFee, 0, 1e18));
+        protocolFee = uint64(bound(protocolFee, 0, 1e18));
         seedAmount = bound(seedAmount, 1e18, 1e36);
         duration = bound(duration, 1 days, 365 days);
 
@@ -333,14 +333,14 @@ contract NodeFuzzTest is BaseTest {
     // invariant 5: withdrawal penalty never exceeds the value of the max swing factor
 
     function test_fuzz_node_swing_price_previewDeposit_matches(
-        uint256 targetReserveRatio,
-        uint256 maxSwingFactor,
+        uint64 targetReserveRatio,
+        uint64 maxSwingFactor,
         uint256 seedAmount,
         uint256 depositAmount,
         uint256 sharesToRedeem
     ) public {
-        targetReserveRatio = bound(targetReserveRatio, 0.01 ether, 0.1 ether); // todo: extend test or hardcode these values
-        maxSwingFactor = bound(maxSwingFactor, 0, 0.1 ether); // todo: extend test or hardcode these values
+        targetReserveRatio = uint64(bound(targetReserveRatio, 0.01 ether, 0.1 ether)); // todo: extend test or hardcode these values
+        maxSwingFactor = uint64(bound(maxSwingFactor, 0, 0.1 ether)); // todo: extend test or hardcode these values
         seedAmount = bound(seedAmount, 1 ether, maxDeposit);
         depositAmount = bound(depositAmount, 1 ether, maxDeposit);
 
@@ -381,13 +381,13 @@ contract NodeFuzzTest is BaseTest {
     }
 
     function test_fuzz_node_swing_price_deposit_never_exceeds_max(
-        uint256 maxSwingFactor,
-        uint256 targetReserveRatio,
+        uint64 maxSwingFactor,
+        uint64 targetReserveRatio,
         uint256 seedAmount,
         uint256 depositAmount
     ) public {
-        maxSwingFactor = bound(maxSwingFactor, 0.01 ether, 0.99 ether);
-        targetReserveRatio = bound(targetReserveRatio, 0.01 ether, 0.99 ether);
+        maxSwingFactor = uint64(bound(maxSwingFactor, 0.01 ether, 0.99 ether));
+        targetReserveRatio = uint64(bound(targetReserveRatio, 0.01 ether, 0.99 ether));
         seedAmount = bound(seedAmount, 1 ether, maxDeposit);
         depositAmount = bound(depositAmount, 1 ether, maxDeposit);
 
@@ -425,13 +425,13 @@ contract NodeFuzzTest is BaseTest {
     }
 
     function test_fuzz_node_swing_price_redeem_never_exceeds_max(
-        uint256 maxSwingFactor,
-        uint256 targetReserveRatio,
+        uint64 maxSwingFactor,
+        uint64 targetReserveRatio,
         uint256 seedAmount,
         uint256 withdrawalAmount
     ) public {
-        maxSwingFactor = bound(maxSwingFactor, 0.01 ether, 0.99 ether);
-        targetReserveRatio = bound(targetReserveRatio, 0.01 ether, 0.99 ether);
+        maxSwingFactor = uint64(bound(maxSwingFactor, 0.01 ether, 0.99 ether));
+        targetReserveRatio = uint64(bound(targetReserveRatio, 0.01 ether, 0.99 ether));
         seedAmount = bound(seedAmount, 100 ether, 1e36);
 
         deal(address(asset), address(user), seedAmount);
